+++ app/src/main/java/com/themoon/y1/adapters/CategoryListAdapter.java
@@ -1,0 +1,75 @@
+package com.themoon.y1.adapters;
+
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AbsListView;
+import android.widget.BaseAdapter;
+import android.widget.Button;
+
+import com.themoon.y1.MainActivity;
+import com.themoon.y1.ThemeManager;
+
+import java.util.List;
+
+public class CategoryListAdapter extends BaseAdapter {
+    private List<String> items;
+    private String type;
+
+    public CategoryListAdapter(List<String> items, String type) {
+        this.items = items;
+        this.type = type;
+    }
+
+    @Override
+    public int getCount() { return items.size(); }
+
+    @Override
+    public Object getItem(int position) { return items.get(position); }
+
+    @Override
+    public long getItemId(int position) { return position; }
+
+    @Override
+    public View getView(final int position, View convertView, ViewGroup parent) {
+        final Button btn;
+
+        if (convertView == null) {
+            btn = MainActivity.instance.createListButton("");
+            btn.setLayoutParams(new AbsListView.LayoutParams(
+                    AbsListView.LayoutParams.MATCH_PARENT,
+                    AbsListView.LayoutParams.WRAP_CONTENT));
+        } else {
+            btn = (Button) convertView;
+        }
+
+        final String name = items.get(position);
+        btn.setText((type.equals("ARTIST") ? "👤 " : "💿 ") + name);
+
+        btn.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View v, boolean hasFocus) {
+                if (hasFocus) {
+                    btn.setBackground(MainActivity.instance.createButtonBackground(ThemeManager.getListButtonFocusedBg()));
+                    btn.setTextColor(ThemeManager.getListButtonFocusedTextColor());
+                    MainActivity.instance.showFastScrollLetter(name);
+                } else {
+                    btn.setBackground(MainActivity.instance.createButtonBackground(ThemeManager.getListButtonNormalBg()));
+                    btn.setTextColor(ThemeManager.getTextColorPrimary());
+                }
+            }
+        });
+
+        btn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                MainActivity.instance.clickFeedback();
+                MainActivity.instance.virtualQueryType = type;
+                MainActivity.instance.virtualQueryValue = name;
+                MainActivity.instance.currentBrowserMode = MainActivity.BROWSER_VIRTUAL_SONGS;
+                MainActivity.instance.buildVirtualSongs();
+            }
+        });
+
+        return btn;
+    }
+}
