+++ app/src/main/java/com/themoon/y1/adapters/SongListAdapter.java
@@ -1,0 +1,71 @@
+package com.themoon.y1.adapters;
+
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AbsListView;
+import android.widget.BaseAdapter;
+import android.widget.Button;
+
+import com.themoon.y1.MainActivity;
+import com.themoon.y1.ThemeManager;
+import com.themoon.y1.models.SongItem;
+
+import java.util.List;
+
+public class SongListAdapter extends BaseAdapter {
+    private List<SongItem> items;
+
+    public SongListAdapter(List<SongItem> items) {
+        this.items = items;
+    }
+
+    @Override
+    public int getCount() { return items.size(); }
+
+    @Override
+    public Object getItem(int position) { return items.get(position); }
+
+    @Override
+    public long getItemId(int position) { return position; }
+
+    @Override
+    public View getView(final int position, View convertView, ViewGroup parent) {
+        final Button btn;
+
+        if (convertView == null) {
+            btn = MainActivity.instance.createListButton("");
+            btn.setLayoutParams(new AbsListView.LayoutParams(
+                    AbsListView.LayoutParams.MATCH_PARENT,
+                    AbsListView.LayoutParams.WRAP_CONTENT));
+        } else {
+            btn = (Button) convertView;
+        }
+
+        final SongItem song = items.get(position);
+        btn.setText("🎵 " + song.title);
+
+        btn.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View v, boolean hasFocus) {
+                if (hasFocus) {
+                    btn.setBackground(MainActivity.instance.createButtonBackground(ThemeManager.getListButtonFocusedBg()));
+                    btn.setTextColor(ThemeManager.getListButtonFocusedTextColor());
+                    MainActivity.instance.showFastScrollLetter(song.title);
+                } else {
+                    btn.setBackground(MainActivity.instance.createButtonBackground(ThemeManager.getListButtonNormalBg()));
+                    btn.setTextColor(ThemeManager.getTextColorPrimary());
+                }
+            }
+        });
+
+        btn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                MainActivity.instance.clickFeedback();
+                MainActivity.instance.playTrackList(MainActivity.instance.virtualSongList, position);
+            }
+        });
+
+        return btn;
+    }
+}
